Move controllerUrlUtils to SPI (#8573)
