+++ pinot-spi/src/main/java/org/apache/pinot/spi/utils/StringUtil.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.utils;
+package org.apache.pinot.spi.utils;
 
 import org.apache.commons.lang.StringUtils;
